support set float value in duty_cycle_percentage

+++ include/pwm_sysfs.h
@@ -34,7 +34,7 @@ int pwm_open(pwm_t *pwm, unsigned int chip, unsigned int channel);
 int pwm_set_period(pwm_t *pwm, uint64_t period_ns);
 int pwm_set_period_hz(pwm_t *pwm, uint32_t hz);
 int pwm_set_duty_cycle(pwm_t *pwm, uint64_t duty_ns);
-int pwm_set_duty_cycle_percentage(pwm_t *pwm, uint8_t percent);
+int pwm_set_duty_cycle_percentage(pwm_t *pwm, float percent);
 int pwm_enable(pwm_t *pwm, bool enable);
 int pwm_close(pwm_t *pwm);
